+++ docassemble_base/docassemble/base/filter.py
@@ -867,7 +867,7 @@ def image_url_string(match, emoji=False, question=None, playground=False):
         if file_info.get('extension', '') in ['png', 'jpg', 'gif', 'svg', 'jpe', 'jpeg']:
             return('<img class="daicon daimageref" style="' + width_string + '" src="' + the_url + '"/>')
         elif file_info['extension'] == 'pdf':
-            image_url = server.url_finder(file_reference, size="screen", page=1, _question=question, ext=file_info.get('extension', None))
+            image_url = server.url_finder(file_reference, size="screen", page=1, _question=question)
             if image_url is None:
                 return ('[ERROR: File reference ' + str(file_reference) + ' cannot be displayed]')
             output = '<a target="_blank" class="daimageref" href="' + the_url + '"><img class="daicon dapdfscreen" style="' + width_string + '" src="' + image_url + '"/></a>'